requirements/desktop.txt
# Core
Django==2.2.5
django-configurations==2.1

# Model Tools
django-model-utils==3.2.0
jsonfield==2.0.2

# Rest apis
djangorestframework==3.10.2
django-filter==2.2.0
coreapi
django-cors-headers
django-environ==0.4.5
requests==2.22.0

# Logging
user-agents>=2.0


# Developer Tools
ipython==7.8.0
autopep8>=1.4.4
django-nose==1.4.6

# Calc engine
numpy>=1.17.0
numpydoc>=0.9.1
pandas>=0.25.0
openpyxl>=2.6.2
sqlalchemy>=1.3.6
xlrd
jsonpickle>=1.2
xarray>=0.12.3
plotly>=4.1.0
bokeh>=1.3.4
matplotlib>=3.1.1

# Serve
whitenoise>=4.1.4
